config-rules/human-iam-users/requirements.txt
aws_cdk.aws_iam==1.96.0
aws_cdk.aws_ssm==1.96.0
aws_cdk.aws_config==1.96.0
aws-cdk.aws-lambda==1.96.0
aws-cdk.aws-cloudformation==1.96.0
cfn-flip==1.2.3
aws-cdk.aws-events==1.96.0
aws-cdk.aws-events-targets==1.96.0
